Second try at a skeleton for the aggregated backend.
